A new simple game I have tried to make.
If you get this notification can you please explain how to make a collision system and how it works because I have no clue. I tried making one myself then by using the one you made in the paddle game. I looked on the internet and people say "don't reinvent the wheel. Use a framework." I do not want to use a framework because I want to learn the vanilla version of the language. In this game I just need it to detect if either of the 2 white squares hits the player.